Fix a bug in the reverse iteration setup
When the :from value is not equal to a key the iteration stops immediately. The reason is that the first element found is larger than :from (because #peek_next is the next key after :from). In this situation we need to skip one element so that we start with the first element smaller than :from instead of the first element larger than :from.